Prioritise Top Languages
