+++ analyser/application/repository_analysis/prioritised_languages.yml
@@ -1,0 +1,25 @@
+languages:
+  - Python
+  - JavaScript
+  - TypeScript
+  - C
+  - C++
+  - Go
+  - Rust
+  - Ruby
+  - PHP
+  - Swift
+  - Kotlin
+  - Scala
+  - Dart
+  - Shell
+  - SQL
+  - HTML
+  - CSS
+  - Java
+  - R
+  - MATLAB
+  - Jupyter Notebook
+  - Markdown
+  - JSON
+  - YAML
